refactor: move hash processor constants to core
