+++ processor/src/aux_table/tests.rs
@@ -1,11 +1,12 @@
 use super::{
-    super::{
-        hasher::{LINEAR_HASH, RETURN_STATE},
-        ExecutionTrace, Operation, Process,
-    },
+    super::{ExecutionTrace, Operation, Process},
     AuxTableTrace,
 };
-use vm_core::{bitwise::BITWISE_OR, Felt, FieldElement, ProgramInputs, AUX_TRACE_RANGE};
+use vm_core::{
+    bitwise::BITWISE_OR,
+    hasher::{LINEAR_HASH, RETURN_STATE},
+    Felt, FieldElement, ProgramInputs, AUX_TRACE_RANGE,
+};
 
 #[test]
 fn hasher_aux_trace() {
